A bunch of minor touch ups
