+++ coderd/rbac/object.go
@@ -124,10 +124,6 @@ var (
 	ResourceLicense = Object{
 		Type: "license",
 	}
-
-	ResourceMetrics = Object{
-		Type: "metrics",
-	}
 )
 
 // Object is used to create objects for authz checks when you have none in
